tweak service template
