+++ service/workflows/service-templates/index.test.ts
@@ -1,0 +1,8 @@
+import { expect, describe, it } from "vitest";
+import * as mainExports from "@template/file-service";
+
+describe("@template/file-service", () => {
+  it("should be defined", () => {
+    expect(mainExports).toBeDefined();
+  });
+});
