+++ service/workflows/service-templates/package.json
@@ -2,6 +2,9 @@
   "name": "@template/file-service",
   "description": "Service package for template-file",
   "type": "module",
+  "exports": {
+    ".": "./index.ts"
+  },
   "scripts": {
     "start": "node --experimental-strip-types ./run.ts",
     "test": "vitest run",
